Moved ruleset service mock to mock package
