+++ api/server/api_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/heetch/regula"
 	"github.com/heetch/regula/api"
+	"github.com/heetch/regula/mock"
 	"github.com/heetch/regula/rule"
 	"github.com/heetch/regula/store"
 	"github.com/pkg/errors"
@@ -21,7 +22,7 @@ import (
 )
 
 func TestAPI(t *testing.T) {
-	s := new(mockRulesetService)
+	s := new(mock.RulesetService)
 	h := NewHandler(s, Config{
 		WatchTimeout: 1 * time.Second,
 	})
@@ -359,7 +360,7 @@ func TestAPI(t *testing.T) {
 	})
 }
 
-func resetStore(s *mockRulesetService) {
+func resetStore(s *mock.RulesetService) {
 	s.ListCount = 0
 	s.LatestCount = 0
 	s.OneByVersionCount = 0
